feat: add support to add polylines
Close #102
