+++ src/main/java/com/flowingcode/vaadin/addons/googlemaps/GoogleMapPolyline.java
@@ -1,0 +1,33 @@
+/*-
+ * #%L
+ * Google Maps Addon
+ * %%
+ * Copyright (C) 2020 - 2024 Flowing Code
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+package com.flowingcode.vaadin.addons.googlemaps;
+
+import java.util.List;
+
+@SuppressWarnings("serial")
+public class GoogleMapPolyline extends GoogleMapPoly {
+
+  public GoogleMapPolyline(List<GoogleMapPoint> points) {
+    super(points);
+    setClosed(false);
+  }  
+
+}
